Refactored Interop and started with host support. VstHostCommandProxy will take on calls from the plugin (audioMasterCallback), VstPluginCommandStub will dispatch calls to the plugin (AEffect funct. ptr's) and VstPluginContext contains all information on a specific plugin instance. The Initialize method will load the plugin dll and call the exported main function.

+++ Source/Code/Jacobi.Vst.Core/IVstHostCommands.cs
@@ -61,7 +61,7 @@
         /// Retrieves the current sample rate from the host.
         /// </summary>
         /// <returns>Returns the number of samples per second.</returns>
-        double GetSampleRate();
+        float GetSampleRate();
         /// <summary>
         /// Retrieves the number of samples passed to the plugin during the audio processing cycles.
         /// </summary>
